Use wireframe placeholder component in other wireframe stories
