+++ packages/onboarding/src/step-container-v2/helpers/withStepContainerV2ContextDecorator.tsx
@@ -9,22 +9,3 @@ export const withStepContainerV2ContextDecorator = ( Story: React.ComponentType 
 		</StepContainerV2Provider>
 	);
 };
-
-export function WireframePlaceholder( {
-	height,
-	children,
-}: {
-	height?: number;
-	children?: React.ReactNode;
-} ) {
-	const style = {
-		background: '#ff80ff',
-		borderRadius: 10,
-		display: 'flex',
-		alignItems: 'center',
-		justifyContent: 'center',
-		...( height && { height } ),
-	};
-
-	return <div style={ style }>{ children }</div>;
-}
