+++ packages/onboarding/src/step-container-v2/wireframes/FullWidthLayout/style.stories.scss
@@ -36,7 +36,6 @@
 
 	&__preview {
 		flex: 1;
-		background: #fa8072;
 	}
 
 	&__fonts {
