+++ packages/onboarding/src/step-container-v2/wireframes/SixColumnsCenteredLayout/style.scss
@@ -2,5 +2,5 @@
 	max-width: 615px;
 	display: flex;
 	flex-direction: column;
-	align-items: center;
+	align-items: stretch;
 }
